Move datasourceversion fetching into custom reusable hook
